requirements.txt
streamlit
chromadb
sentence-transformers
langchain
langchain-community
python-dotenv
pypdf
python-docx
tqdm
requests
langdetect
